feat(toggleterm): only create terminals once
Before, a new terminal instance would be created each time the keymap is
invoked. This means that terminals were not really being toggled, which
not only means we're not saving the buffer contents, but we're also
wasting resources.

Now, each terminal is created with a unique identifier, which allows us
to truly toggle it and instead of closing an open terminal, we're now
hiding it, making the toggling process much faster and efficient.